feat: Implement caching for currency rates, enhance error handling, and improve request logging

+++ server/src/main/java/com/taskxcode/task_xcode/dto/CurrencyResponse.java
@@ -1,21 +1,23 @@
 package com.taskxcode.task_xcode.dto;
 
+import java.math.BigDecimal;
+
 public class CurrencyResponse {
 
-    private double value;
+    private BigDecimal value;
 
     public CurrencyResponse() {
     }
 
-    public CurrencyResponse(double value) {
+    public CurrencyResponse(BigDecimal value) {
         this.value = value;
     }
 
-    public double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
 }
